+++ packages/vm/src/instance.rs
@@ -393,7 +393,7 @@ mod singlepass_test {
 
         let init_used = orig_gas - instance.get_gas();
         println!("init used: {}", init_used);
-        assert_eq!(init_used, 45568);
+        assert_eq!(init_used, 65606);
     }
 
     #[test]
@@ -417,7 +417,7 @@ mod singlepass_test {
 
         let handle_used = gas_before_handle - instance.get_gas();
         println!("handle used: {}", handle_used);
-        assert_eq!(handle_used, 63432);
+        assert_eq!(handle_used, 94480);
     }
 
     #[test]
@@ -452,6 +452,6 @@ mod singlepass_test {
 
         let query_used = gas_before_query - instance.get_gas();
         println!("query used: {}", query_used);
-        assert_eq!(query_used, 23066);
+        assert_eq!(query_used, 32676);
     }
 }